requirements.txt
absl-py==1.3.0
anyio==3.6.2
astunparse==1.6.3
boto3==1.26.28
botocore==1.29.28
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
fastapi==0.88.0
flatbuffers==22.12.6
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h11==0.14.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.1.0
jmespath==1.0.1
keras==2.11.0
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
mongoengine==0.24.2
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
google-api-python-client==2.108.0